src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.34.0
python-multipart==0.0.19
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==24.1.0
langchain==0.1.20
PyMuPDF==1.24.3
motor==3.6.0
openai==1.55.3
qdrant-client==1.10.1
passlib[bcrypt]==1.7.4
